As of Jan 26 2003 the -W label is dead. See Revision 1.67 http://www.freebsd.org/cgi/cvsweb.cgi/src/sbin/disklabel/disklabel.c?rev=1.67&content-type=text/x-cvsweb-markup
Retire the "-N" and "-W" options which implemented write-protection
for the disklabel:  This facility is OBE.

First of all, we cannot sensibly implement this in a properly stacked
environment.

Second, if we did, it would confuse the heck out of users who
wouldn't be able to "start from scratch" by dd(8)'ing /dev/zero
onto /dev/da0.

Third, the offered protection is not comprehensive: no other software
would respect it.

Fourth and finally, the disklabel is already protected against
tampering if it controls open partitions.

Uselessness of these options discussed with:    peter