Alternative heatmap generation 'framework'. Simple/Single file with no persistent storage, making debugging easy.
This proves that the heatmap code itself is not causing the issues, but the
image manipulation PIL is not so fast (or properly utilized). Might want to
consider a more advanced image generation toolkit. (2000 entries: 5+ secs).

Also loading a large dataset into Objects seems to be very slow. The database
call is effient (just a single one), but building all the related objects,
makes is slow (2000 entries: 1+ sec). Might want to check if raw queries suit
to avoid conversion our needs:
   http://docs.djangoproject.com/en/dev/topics/db/sql/