I guess the python documentation is right about this one :-)
"Lots of people want their programs to have “required options”. Think about it.
If it’s required, then it’s not optional! If there is a piece of information
that your program absolutely requires in order to run successfully, that’s what
positional arguments are for." - http://docs.python.org/library/optparse.html

Fixed.