Control gateway routing using 'flagged' IP.
The use of specialized serviceid IP addresses for use in determining the
presence of a default gateway will allow us, to remove the IP if the gateway is
not functional, without affecting other functions of the node. The presence of
the unused entries allowes new gateways to be added without the need of
updating all configuration files.


Side-note: This is a workaround since default routing to pasted through
lvrouted using the '-g' parameter at gateways and leaving the '-z' parameter
empty other nodes. How-ever this setup is causing defaultgw 'loops' in the
network, yet need to be tested some more.