Implement named replacement for FreeBSD 11.
Unbound via ports, configured to listen to only specific interfaces.
For security purposes do not listen to outside world.

Depending on it's location, making sure the forwarders are internal 
or external.

Note: Dropping support for autoritive nameserver, this feature is not 
required anymore since our network is running stable now.