Speed-up image transfer, by using compression.
The 400MB image could be compressed to roughly 123MB (4:1 compression ratio)
when doing file based gzip compression. Yet, since we cannot store the
temponary file at the node we are stuck in compressing it on-the-fly. Still a
nice optimalisation from 144 sec to 87 sec (on a wired connection). It could
even be faster but the limiting factor currently is the CPU-power on the board
for decompression and next the write-speed to the CF disk.

When using a 10Mbit connection I estimate roughly 300 sec for an update to take
place.