Try to beat captive portal requests for already autorized devices.
Dummy framework for tool which reads arp requests and based on the response
checks if the MAC is already whitelisted and hence adds it's IP to the
whitelist table. 

This will avoid issues with captive portal checkers at phone, which assume a
captive portal, yet their request in and on itself will already activate the
known lease again. 

Secondly this framework is also required to allow 'roaming' from users between
accesspoints without the need of re-authentication (or IoT devices for
example).