Update OnsBuiten unified SSID requiries bridging
If all SSIDs are marked equally (in this case 'WiFi Leiden') the interfaces
needs to be bridged, else a moving client from AP1 to AP2 will not be able to
reach the gateway anymore.

The configured setup with 3 subnets is actually overkill, yet left in place as
example on how-to convert an node without reconfiguring it's device IPs as one
could use a single /24 range for all devices.