Fix adding of interface members of bridge(4)
create_args_ is not supported for bridge creation, how-ever when using
ifconfig_ some trickery is required to support IP assignments as all needs to
be 'shifted' to aliasX interfacing.

Note alias creation is flawed by design as they all have to have unique keys,
thus we rely on _aliasX definitions for alias creation, where-as we should
parse the entries instead. Thus make_interface_list has to hack support for
bridge aliases which do require this tooling.