Add PoC for persistent masterip range
Use 1 IP from the PoC defined masterip range 172.31.254.0/24 as alias on lo0
this will ensure the IP will always exist and is persistent. 

This is quite usefull when connecting for example with SSH to the masterip and
doing maintenance on this attached interface at the same time, instead of
shooting yourself in the foot by bringing down the masterip you are connecting
to.

Added bonus; scanning for available nodes is more easy by just sweeping one
range.