Fix alias VLAN dhcp declaration not mapped to shared-network
While here make all entries-shared networks, since it does not harm anyways:

Note that even when the shared-network declaration is absent, an  empty
one  is  created  by  the  server to contain the subnet (and any scoped
parameters included in the subnet).  For practical purposes, this means
that  "stateless"  DHCP  clients,  which are not tied to addresses (and
therefore subnets) will receive  the  same  configuration  as  stateful
ones.