Fix no traffic possible from aliased gateway iface
Packets originating from same interface as the gateway interface (aliased
interface) are not matched by the 'out' rule. 

Technically they are both incoming and outgoing at the same time and thus will
be covered the the 'in' rule directive. By ensuring matching is done on both
incoming and outgoing this packets will be allowed.